fix: change the layout of ts interface, types & enums
