+++ frontend/src/components/AuthTitle/IAuthTitle.ts
@@ -1,0 +1,3 @@
+export interface IAuthTitleProps {
+  titleText: string;
+}
