+++ frontend/src/components/Avatar/Avatar.tsx
@@ -2,13 +2,9 @@ import { FC } from "react";
 import "./Avatar.scss";
 import avatar from "../../images/avatar.svg";
 import { NavLink } from "react-router-dom";
+import { IAvatarProps } from "./IAvatar";
 
-interface AvatarProps {
-  handleCloseBurgerMenu?: () => void;
-  isBurgerOpen?: boolean;
-}
-
-const Avatar: FC<AvatarProps> = ({ handleCloseBurgerMenu, isBurgerOpen }) => {
+const Avatar: FC<IAvatarProps> = ({ handleCloseBurgerMenu, isBurgerOpen }) => {
   return (
     <NavLink className="avatar" onClick={handleCloseBurgerMenu} to="/profile">
       <img className="avatar-image" src={avatar} alt="Аватар" />
