+++ frontend/src/components/Avatar/IAvatar.ts
@@ -1,0 +1,4 @@
+export interface IAvatarProps {
+  handleCloseBurgerMenu?: () => void;
+  isBurgerOpen?: boolean;
+}
