+++ frontend/src/components/Form/IForm.ts
@@ -1,0 +1,12 @@
+import { FormParams } from "../../utils/Api/ApiTypes";
+
+export interface IFormProps {
+  buttonText: string;
+  handleSubmitForm: (values: FormParams) => void;
+}
+
+export type FormInputs = {
+  login: string;
+  email: string;
+  password: string;
+};
