+++ frontend/src/components/Guide/IGuide.ts
@@ -1,0 +1,10 @@
+export interface IIcon {
+  className: string;
+  src: string;
+  alt: string;
+}
+
+export interface IDetails {
+  text: string[];
+  icons: IIcon[];
+}
