+++ frontend/src/components/HOC/IProtectedRoute.ts
@@ -1,0 +1,4 @@
+export interface IProtectedRouteProps {
+  loggedIn: boolean;
+  children: any;
+}
