+++ frontend/src/components/HOC/ProtectedRoute.tsx
@@ -1,12 +1,8 @@
 import { FC } from "react";
 import { Navigate } from "react-router-dom";
+import { IProtectedRouteProps } from "./IProtectedRoute";
 
-interface ProtectedRouteProps {
-  loggedIn: boolean;
-  children: any;
-}
-
-const ProtectedRoute: FC<ProtectedRouteProps> = ({ loggedIn, children }) => {
+const ProtectedRoute: FC<IProtectedRouteProps> = ({ loggedIn, children }) => {
   if (!loggedIn) {
     return <Navigate to="/" />;
   }
