+++ frontend/src/components/Header/BurgerButton/IBurgerButton.ts
@@ -1,0 +1,5 @@
+export interface IBurgerButtonProps {
+  handleOpenBurgerMenu: () => void;
+  handleCloseBurgerMenu: () => void;
+  isBurgerOpen: boolean;
+}
