+++ frontend/src/components/Header/BurgerMenu/IBurgerMenu.ts
@@ -1,0 +1,4 @@
+export interface IBurgerMenuProps {
+  handleCloseBurgerMenu: () => void;
+  isBurgerOpen: boolean;
+}
