+++ frontend/src/components/Header/IHeader.ts
@@ -1,0 +1,3 @@
+export interface IHeaderProps {
+  loggedIn: boolean;
+}
