+++ frontend/src/components/Header/Nav/INav.ts
@@ -1,0 +1,4 @@
+export interface INavProps {
+  handleCloseBurgerMenu?: () => void;
+  isBurgerOpen?: boolean;
+}
