+++ frontend/src/components/Header/Nav/Nav.tsx
@@ -2,15 +2,11 @@ import { FC } from "react";
 import { NavLink } from "react-router-dom";
 import "./Nav.scss";
 import { header } from "../../../utils/textConstants";
+import { INavProps } from "./INav";
 
 const setActive = ({ isActive }: any) => `nav__link ${isActive && "nav__link_active"}`;
 
-interface NavProps {
-  handleCloseBurgerMenu?: () => void;
-  isBurgerOpen?: boolean;
-}
-
-const Nav: FC<NavProps> = ({ handleCloseBurgerMenu, isBurgerOpen }) => {
+const Nav: FC<INavProps> = ({ handleCloseBurgerMenu, isBurgerOpen }) => {
   return (
     <section className={`nav ${isBurgerOpen && "nav_burger"}`}>
       {isBurgerOpen && (
