+++ frontend/src/components/Login/ILogin.ts
@@ -1,0 +1,7 @@
+import { SetStateAction } from "react";
+
+export interface ILoginProps {
+  handleAuthorize: (values: any) => void;
+  errorMesage: string;
+  setErrorMesage: React.Dispatch<SetStateAction<string>>;
+}
