+++ frontend/src/components/Login/Login.tsx
@@ -1,18 +1,14 @@
-import { FC, SetStateAction } from "react";
+import { FC } from "react";
 import AuthTitle from "../AuthTitle/AuthTitle";
 import Form from "../Form/Form";
 import "./Login.scss";
 import Popup from "../Popup/Popup";
 import { login } from "../../utils/textConstants";
-import Reference, { LinkVariant } from "../Reference/Reference";
+import Reference from "../Reference/Reference";
+import { ILoginProps } from "./ILogin";
+import { LinkVariant } from "../Reference/IReference";
 
-interface LoginProps {
-  handleAuthorize: (values: any) => void;
-  errorMesage: string;
-  setErrorMesage: React.Dispatch<SetStateAction<string>>;
-}
-
-const Login: FC<LoginProps> = ({ handleAuthorize, errorMesage, setErrorMesage }) => {
+const Login: FC<ILoginProps> = ({ handleAuthorize, errorMesage, setErrorMesage }) => {
   const closePopup = (evt: any) => {
     if (
       evt.currentTarget === evt.target ||
