+++ frontend/src/components/Main/Greeting/IGreeting.ts
@@ -1,0 +1,5 @@
+export interface IGreetingProps {
+  greetingTitle: string;
+  greetingSubtitle: string;
+  buttonTitle: string;
+}
