+++ frontend/src/components/Main/IMain.ts
@@ -1,0 +1,7 @@
+import { SetStateAction } from "react";
+
+export interface IMainProps {
+  handleFeedback: (values: any) => void;
+  errorMesage: string;
+  setErrorMesage: React.Dispatch<SetStateAction<string>>;
+}
