+++ frontend/src/components/Popup/IPopup.ts
@@ -1,0 +1,4 @@
+export interface IPopupProps {
+  text: string;
+  closePopup: (evt: any) => void;
+}
