+++ frontend/src/components/Profile/IProfile.ts
@@ -1,0 +1,3 @@
+export interface IProfileProps {
+  handleloggedOutClick: () => void;
+}
