+++ frontend/src/components/Reference/IReference.ts
@@ -1,0 +1,11 @@
+export interface IReferenceProps {
+  textDescription: string;
+  textLink: string;
+  path: LinkVariant;
+}
+
+export enum LinkVariant {
+  toRegister = "/signup",
+  toLogin = "/signin",
+  toFeedback = "/#about",
+}
