+++ frontend/src/components/Reference/Reference.tsx
@@ -2,20 +2,9 @@ import { FC } from "react";
 import "./Reference.scss";
 // import { Link } from "react-router-dom";
 import { HashLink as Link } from "react-router-hash-link";
+import { IReferenceProps, LinkVariant } from "./IReference";
 
-export enum LinkVariant {
-  toRegister = "/signup",
-  toLogin = "/signin",
-  toFeedback = "/#about",
-}
-
-interface ReferenceProps {
-  textDescription: string;
-  textLink: string;
-  path: LinkVariant;
-}
-
-const Reference: FC<ReferenceProps> = ({ textDescription, textLink, path }) => {
+const Reference: FC<IReferenceProps> = ({ textDescription, textLink, path }) => {
   return (
     <p className="reference" style={path === LinkVariant.toFeedback ? { marginTop: 100 } : {}}>
       {textDescription}
