+++ frontend/src/components/Register/IRegister.ts
@@ -1,0 +1,7 @@
+import { SetStateAction } from "react";
+
+export interface IRegisterProps {
+  handleRegister: (values: any) => void;
+  errorMesage: string;
+  setErrorMesage: React.Dispatch<SetStateAction<string>>;
+}
