+++ frontend/src/components/Register/Register.tsx
@@ -1,18 +1,14 @@
-import { FC, SetStateAction } from "react";
+import { FC } from "react";
 import AuthTitle from "../AuthTitle/AuthTitle";
 import Form from "../Form/Form";
 import "./Register.scss";
 import Popup from "../Popup/Popup";
 import { register } from "../../utils/textConstants";
-import Reference, { LinkVariant } from "../Reference/Reference";
+import Reference from "../Reference/Reference";
+import { LinkVariant } from "../Reference/IReference";
+import { IRegisterProps } from "./IRegister";
 
-interface RegisterProps {
-  handleRegister: (values: any) => void;
-  errorMesage: string;
-  setErrorMesage: React.Dispatch<SetStateAction<string>>;
-}
-
-const Register: FC<RegisterProps> = ({ handleRegister, errorMesage, setErrorMesage }) => {
+const Register: FC<IRegisterProps> = ({ handleRegister, errorMesage, setErrorMesage }) => {
   const closePopup = (evt: any) => {
     if (
       evt.currentTarget === evt.target ||
