+++ frontend/src/utils/Api/ApiTypes.ts
@@ -1,0 +1,11 @@
+export type FeedbackParams = {
+  title: string;
+  return_address: string;
+  message: string;
+};
+
+export type FormParams = {
+  username: string;
+  email: string;
+  password: string;
+};
